refactor: use domain as constant and delete lib/constants
