+++ api/graduates.ts
@@ -1,3 +1,4 @@
+import { domain } from 'util/const';
 interface GraduatesAPIData {
   id: number;
   session: number;
@@ -16,7 +17,7 @@ export interface Graduate {
 }
 
 export const getGraduates = async (session: number): Promise<Graduate[]> => {
-  const response = await fetch(`${process.env.NEXT_PUBLIC_API_ORIGIN}/graduates?session=${session}`);
+  const response = await fetch(`${domain}/graduates?session=${session}`);
   const data: GraduatesAPIData[] = await response.json();
 
   if (!data.length) {
