+++ api/seos.ts
@@ -1,8 +1,8 @@
-import { DEFAULT_META } from 'lib/constants';
+import { DEFAULT_META, domain } from 'util/const';
 import { Meta } from 'components/atoms/PageMeta/PageMeta';
 
 export const getSeos = async (path: string): Promise<Meta> => {
-  const response = await fetch(`${process.env.NEXT_PUBLIC_API_ORIGIN}/seos?path=${path}`);
+  const response = await fetch(`${domain}/seos?path=${path}`);
   const data = await response.json();
 
   if (!data.length) {
