+++ api/visitations.ts
@@ -1,4 +1,4 @@
-const domain = process.env.NEXT_PUBLIC_API_ORIGIN;
+import { domain } from 'util/const';
 
 export interface Visitation {
   id: number;
